examples/python/sui-demo/pyproject.toml
[project]
name = "sui-demo"
version = "1.0.0"
description = "Simple SUI X402 payment demo with host and merchant agents"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.0",
    "uvicorn>=0.34.2",
    "python-dotenv",
    "starlette>=0.47.3",
    "a2a-sdk>=0.3.3",
    "google-adk>=1.12.0",
    "x402[sui] @ file:///Users/michaelhayes/x402/python",
    "a2a-x402",
    "fastapi",
    "httpx",
    "pysui>=0.88.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["client", "server"]

[tool.uv]
override-dependencies = [
    "httpx==0.27.2",
    "websockets==12.0",
]

[tool.uv.sources]
x402 = { path = "/Users/michaelhayes/x402/python/x402", editable = true }
a2a-x402 = { path = "../../../python/a2a_x402", editable = true }
